Move database items into specific namespace
